+++ NEWS.md
@@ -1,3 +1,7 @@
+## Version `v1.0.1` (in progress)
+* Folio secure store proxy deployment issues (FSSP-6)
+---
+
 ## Version `v1.0.0` (31.10.2025)
 * Develop a Dedicated Service as a Secure Key Store (FSSP-1)
 * Implement mTLS Server-Side Authentication (FSSP-4)